Move StoryDb to storage package.
