+++ app/src/main/java/lt/gyvosistorijos/SyncController.kt
@@ -7,6 +7,7 @@ import com.bluelinelabs.conductor.Controller
 import com.bluelinelabs.conductor.RouterTransaction
 import kotlinx.android.synthetic.main.controller_sync.view.*
 import lt.gyvosistorijos.entity.Story
+import lt.gyvosistorijos.storage.StoryDb
 import lt.gyvosistorijos.utils.AppEvent
 import retrofit2.Call
 import retrofit2.Callback
