+++ app/src/main/java/lt/gyvosistorijos/ui/onboarding/OnboardingIntroController.kt
@@ -9,7 +9,7 @@ import com.google.android.gms.maps.GoogleMap
 import kotlinx.android.synthetic.main.controller_onboarding_intro.view.*
 import lt.gyvosistorijos.MainActivity
 import lt.gyvosistorijos.R
-import lt.gyvosistorijos.StoryDb
+import lt.gyvosistorijos.storage.StoryDb
 import lt.gyvosistorijos.utils.AppEvent
 import lt.gyvosistorijos.utils.addTaggedStoryMarkers
